reorganise a little
